shopping cart has been fully implemented
The shopping cart is completely functional and has all the functionality of the template site. it only has one bug and that's the same bug that the template site has. And that's that hitting refresh after adding an item to the cart ads that same amount again. i will try and work out a solution